Revert "Revert "member avatars,editList, createCard etc.""
This reverts commit 5393e01e1c779a427759e757da05c7a89136903e.

+++ src/components/list/index.js
@@ -98,17 +98,19 @@ export default function List(props) {
                     ref={dropdownRef}
                     className={`${styles.menu} ${isActive ? styles.active : ''}`}
                 >
-                    <div>
-                        <ButtonClean
+                    <div onClick={() => setIsVisibleEdit(!isVisibleEdit)}>
+                        {/* <ButtonClean
                             onClick={() => setIsVisibleEdit(!isVisibleEdit)}
                             title='Edit'
-                        />
+                        /> */}
+                        Edit
                     </div>
-                    <div>
-                        <ButtonClean
+                    <div onClick={() => { if (window.confirm('Are you sure you wish to delete this item?')) deleteList() }}>
+                        {/* <ButtonClean
                             onClick={() => { if (window.confirm('Are you sure you wish to delete this item?')) deleteList() }}
                             title='Delete'
-                        />
+                        /> */}
+                        Delete
                     </div>
                 </div>
             </div>
@@ -153,7 +155,7 @@ export default function List(props) {
                 {
                     isVisible ?
                         <form ref={cardRef} className={styles.container} >
-                            <input className={styles.input} type={'text'} value={cardName} onChange={e => setCardName(e.target.value)} />
+                            <input className={styles.taskInput} type={'text'} value={cardName} onChange={e => setCardName(e.target.value)} />
                             <ButtonClean type='submit' className={styles.addlist} onClick={addCard} title='+ Add Task' project={props.project} />
                         </form> : <ButtonClean className={styles['add-task']} onClick={() => setIsVisible(!isVisible)} title='+ Add Task' />
